raspberrypi_code/requirements.txt
Adafruit-Blinka==8.56.0
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-neopixel==6.3.15
adafruit-circuitpython-pixelbuf==2.0.7
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
beautifulsoup4==4.13.3
binho-host-adapter==0.1.6
certifi==2025.1.31
charset-normalizer==3.4.1
docopt==0.6.2
idna==3.10
Js2Py==0.74
packaging==24.2
pipwin==0.5.2
pyftdi==0.56.0
pyjsparser==2.7.1
PyPrind==2.11.3
pyserial==3.5
pySmartDL==1.3.4
pyusb==1.3.1
requests==2.32.3
RPi.GPIO==0.7.1
six==1.17.0
soupsieve==2.6
spidev==3.6
sysv_ipc==1.1.0
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.3.0
